+++ lib/open_weather/raise_error.rb
@@ -6,10 +6,10 @@ module OpenWeather
       def on_complete(env)
         case env[:status]
         when 404
-          raise Faraday::Error::ResourceNotFound, response_values(env)
+          raise Faraday::ResourceNotFound, response_values(env)
         when 407
           # mimic the behavior that we get with proxy requests with HTTPS
-          raise Faraday::Error::ConnectionFailed, %(407 "Proxy Authentication Required ")
+          raise Faraday::ConnectionFailed, %(407 "Proxy Authentication Required ")
         when (400...600).freeze
           raise OpenWeather::Errors::Fault, response_values(env)
         end