+++ spec/open_weather/errors/not_found_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe 'on not found' do
+  include_context 'API client'
+
+  it 'raises error', vcr: { cassette_name: 'errors/not_found' } do
+    expect { client.current_weather(city: 'London') }.to raise_error Faraday::ResourceNotFound
+  end
+end