+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-### 0.1.1 (Next)
+### 0.2.0 (Next)
 
-* [#12](https://github.com/dblock/open-weather-ruby-client/pull/12): Cache Faraday::Connection for persistent adapters - [@dblock](https://github.com/dblock).
+* [#12](https://github.com/dblock/open-weather-ruby-client/pull/12): Cache `Faraday::Connection` for persistent adapters - [@dblock](https://github.com/dblock).
+* [#13](https://github.com/dblock/open-weather-ruby-client/pull/13): Require Faraday >= 1.0 - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2020/05/01)